rename LabelContainer -> TextLabels, move to textLabel.h
