+++ core/src/text/fontContext.h
@@ -22,9 +22,6 @@ constexpr int maxTextures = 64;
 
 namespace alf = alfons;
 
-struct FontMetrics {
-    float ascender, descender, lineHeight;
-};
 
 // TODO could be a shared_ptr<Texture>
 struct GlyphBatch {
